update flightlines, add tiles version
