+++ images/py-figs/NEON_AOP_Hyperspectral_Functions_Tiles_py/linear_contrast.gif
