+++ images/py-figs/NEON_AOP_Hyperspectral_Functions_Tiles_py/output_16_0.png
