+++ images/py-figs/NEON_AOP_Hyperspectral_Functions_Tiles_py/output_30_0.png
