+++ tutorials-in-development/DI-remote-sensing-python/Day1_Hyperspectral_Intro/2018_1.2_Hyperspectral_Functions_RGB_Images.pdf
