--- tutorials-in-development/DI-remote-sensing-python/Day1_Hyperspectral_Intro/Day1_Lesson3_Plot_Spectral_Signature/notebook/neon_aop_hyperspectral.py
@@ -1,157 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Wed Jun 20 10:34:49 2018
-
-@author: bhass
-"""
-
-import matplotlib.pyplot as plt
-import numpy as np
-import h5py, os, copy
-
-def aop_h5refl2array(refl_filename):
-    """aop_h5refl2array reads in a NEON AOP reflectance hdf5 file and returns 
-           1. reflectance array (with the no data value and reflectance scale factor applied)
-           2. dictionary of metadata including spatial information, and wavelengths of the bands
-    --------
-    Parameters
-        refl_filename -- full or relative path and name of reflectance hdf5 file
-    --------
-    Returns 
-    --------
-    reflArray:
-        array of reflectance values
-    metadata:
-        dictionary containing the following metadata:
-            bad_band_window1 (tuple)
-            bad_band_window2 (tuple)
-            bands: # of bands (float)
-            data ignore value: value corresponding to no data (float)
-            epsg: coordinate system code (float)
-            map info: coordinate system, datum & ellipsoid, pixel dimensions, and origin coordinates (string)
-            reflectance scale factor: factor by which reflectance is scaled (float)
-            wavelength: wavelength values (float)
-            wavelength unit: 'm' (string)
-    --------
-    NOTE: This function applies to the NEON hdf5 format implemented in 2016, and should be used for
-    data acquired 2016 and after. Data in earlier NEON hdf5 format (collected prior to 2016) is 
-    expected to be re-processed after the 2018 flight season. 
-    --------
-    Example Execution:
-    --------
-    sercRefl, sercRefl_metadata = h5refl2array('NEON_D02_SERC_DP3_368000_4306000_reflectance.h5') """
-    
-    import h5py
-    
-    #Read in reflectance hdf5 file 
-    hdf5_file = h5py.File(refl_filename,'r')
-
-    #Get the site name
-    file_attrs_string = str(list(hdf5_file.items()))
-    file_attrs_string_split = file_attrs_string.split("'")
-    sitename = file_attrs_string_split[1]
-    
-    #Extract the reflectance & wavelength datasets
-    refl = hdf5_file[sitename]['Reflectance']
-    reflData = refl['Reflectance_Data']
-    reflRaw = refl['Reflectance_Data'].value
-    
-    #Create dictionary containing relevant metadata information
-    metadata = {}
-    metadata['map info'] = refl['Metadata']['Coordinate_System']['Map_Info'].value
-    metadata['wavelength'] = refl['Metadata']['Spectral_Data']['Wavelength'].value
-
-    #Extract no data value & scale factor
-    metadata['data ignore value'] = float(reflData.attrs['Data_Ignore_Value'])
-    metadata['reflectance scale factor'] = float(reflData.attrs['Scale_Factor'])
-    #metadata['interleave'] = reflData.attrs['Interleave']
-    
-    #Apply no data value
-    reflClean = reflRaw.astype(float)
-    arr_size = reflClean.shape
-    if metadata['data ignore value'] in reflRaw:
-        print('% No Data: ',np.round(np.count_nonzero(reflClean==metadata['data ignore value'])*100/(arr_size[0]*arr_size[1]*arr_size[2]),1))
-        nodata_ind = np.where(reflClean==metadata['data ignore value'])
-        reflClean[nodata_ind]=np.nan 
-    
-    #Apply scale factor
-    reflArray = reflClean/metadata['reflectance scale factor']
-    
-    #Extract spatial extent from attributes
-    metadata['spatial extent'] = reflData.attrs['Spatial_Extent_meters']
-    
-    #Extract bad band windows
-    metadata['bad band window1'] = (refl.attrs['Band_Window_1_Nanometers'])
-    metadata['bad band window2'] = (refl.attrs['Band_Window_2_Nanometers'])
-    
-    #Extract projection information
-    #metadata['projection'] = refl['Metadata']['Coordinate_System']['Proj4'].value
-    metadata['epsg'] = int(refl['Metadata']['Coordinate_System']['EPSG Code'].value)
-    
-    #Extract map information: spatial extent & resolution (pixel size)
-    mapInfo = refl['Metadata']['Coordinate_System']['Map_Info'].value
-    
-    hdf5_file.close        
-    
-    return reflArray, metadata
-
-def plot_aop_refl(band_array,refl_extent,colorlimit=(0,1),ax=plt.gca(),title='',cbar ='on',cmap_title='',colormap='Greys'):
-    
-    '''plot_refl_data reads in and plots a single band or 3 stacked bands of a reflectance array
-    --------
-    Parameters
-    --------
-        band_array: array of reflectance values, created from aop_h5refl2array
-        refl_extent: extent of reflectance data to be plotted (xMin, xMax, yMin, yMax) 
-                     use metadata['spatial extent'] from aop_h5refl2array function
-        colorlimit: optional, range of values to plot (min,max). 
-                    - helpful to look at the histogram of reflectance values before plotting to determine colorlimit.
-        ax: optional, default = current axis
-        title: optional; plot title (string)
-        cmap_title: optional; colorbar title 
-        colormap: optional (string, see https://matplotlib.org/examples/color/colormaps_reference.html) for list of colormaps
-    --------
-    Returns 
-    --------
-        plots flightline array of single band of reflectance data
-    --------
-
-    Examples:
-    --------
-    plot_aop_refl(sercb56,
-              sercMetadata['spatial extent'],
-              colorlimit=(0,0.3),
-              title='SERC Band 56 Reflectance',
-              cmap_title='Reflectance',
-              colormap='Greys_r') '''
-    
-    import matplotlib.pyplot as plt
-    
-    plot = plt.imshow(band_array,extent=refl_extent,clim=colorlimit); 
-    if cbar == 'on':
-        cbar = plt.colorbar(plot,aspect=40); plt.set_cmap(colormap); 
-        cbar.set_label(cmap_title,rotation=90,labelpad=20)
-    plt.title(title); ax = plt.gca(); 
-    ax.ticklabel_format(useOffset=False, style='plain'); #do not use scientific notation for ticklabels
-    rotatexlabels = plt.setp(ax.get_xticklabels(),rotation=90); #rotate x tick labels 90 degrees
-    
-def stack_rgb(reflArray,bands):
-    
-    red = reflArray[:,:,bands[0]-1]
-    green = reflArray[:,:,bands[1]-1]
-    blue = reflArray[:,:,bands[2]-1]
-    
-    stackedRGB = np.stack((red,green,blue),axis=2)
-    
-    return stackedRGB
-
-def plot_aop_rgb(rgbArray,ext,ls_pct=5,plot_title=''):
-    
-    from skimage import exposure
-    
-    pLow, pHigh = np.percentile(rgbArray[~np.isnan(rgbArray)], (ls_pct,100-ls_pct))
-    img_rescale = exposure.rescale_intensity(rgbArray, in_range=(pLow,pHigh))
-    plt.imshow(img_rescale,extent=ext)
-    plt.title(plot_title + '\n Linear ' + str(ls_pct) + '% Contrast Stretch'); 
-    ax = plt.gca(); ax.ticklabel_format(useOffset=False, style='plain') #do not use scientific notation #
-    rotatexlabels = plt.setp(ax.get_xticklabels(),rotation=90) #rotate x tick labels 90 degree
