--- tutorials-in-development/DI-remote-sensing-python/Day1_Hyperspectral_Intro/README.txt
@@ -1,1 +1,0 @@
-Updated intro hyperspectral lessons to work with tiles, update functions, cut out redundant/unnecessary sections, etc.
