+++ tutorials-in-development/DI-remote-sensing-python/neon_aop_spectral_python_functions/linearStretch.py
@@ -1,0 +1,10 @@
+def linearStretch(percent):
+    pLow, pHigh = np.percentile(rgbArray[~np.isnan(rgbArray)], (percent,100-percent))
+    img_rescale = exposure.rescale_intensity(rgbArray, in_range=(pLow,pHigh))
+    plt.imshow(img_rescale,extent=sercMetadata['spatial extent'])
+    plt.title('SERC RGB \n Linear ' + str(percent) + '% Contrast Stretch'); 
+    ax = plt.gca()
+    ax.ticklabel_format(useOffset=False, style='plain') 
+    rotatexlabels = plt.setp(ax.get_xticklabels(),rotation=90) 
+
+interact(linearStretch,percent=(0,20,1))
